+++ portal/src/main/java/com/hellblazer/luciferase/portal/mesh/explorer/Abstract3DApp.java
@@ -14,7 +14,9 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.hellblazer.luciferase.portal;
+package com.hellblazer.luciferase.portal.mesh.explorer;
+
+import com.hellblazer.luciferase.portal.Xform;
 
 import javafx.application.Application;
 import javafx.event.EventHandler;
