+++ portal/src/main/java/com/hellblazer/luciferase/portal/mesh/explorer/GraphicsView.java
@@ -1,4 +1,4 @@
-package com.hellblazer.luciferase.portal;
+package com.hellblazer.luciferase.portal.mesh.explorer;
 
 import java.util.List;
 import java.util.stream.Collectors;
