+++ portal/src/main/java/com/hellblazer/luciferase/portal/mesh/explorer/TestCases.java
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.hellblazer.luciferase.portal;
+package com.hellblazer.luciferase.portal.mesh.explorer;
 
 import java.util.ArrayList;
